destroy db client on sigterm

+++ src/functions/avl-processor/index.ts
@@ -158,7 +158,14 @@ export const handler: SQSHandler = async (event, context) => {
         }
 
         throw e;
-    } finally {
-        await dbClient.destroy();
     }
 };
+
+process.on("SIGTERM", async () => {
+    if (dbClient) {
+        logger.info("Destroying DB client...");
+        await dbClient.destroy();
+    }
+
+    process.exit(0);
+});
