+++ src/functions/avl-tfl-location-retriever/index.ts
@@ -89,7 +89,14 @@ export const handler: Handler = async (event, context) => {
         }
 
         throw error;
-    } finally {
-        await dbClient.destroy();
     }
 };
+
+process.on("SIGTERM", async () => {
+    if (dbClient) {
+        logger.info("Destroying DB client...");
+        await dbClient.destroy();
+    }
+
+    process.exit(0);
+});
