+++ src/functions/table-renamer/index.ts
@@ -81,7 +81,14 @@ export const handler: Handler = async (event, context) => {
         }
 
         throw e;
-    } finally {
-        await dbClient.destroy();
     }
 };
+
+process.on("SIGTERM", async () => {
+    if (dbClient) {
+        logger.info("Destroying DB client...");
+        await dbClient.destroy();
+    }
+
+    process.exit(0);
+});
